chore(parser): move getEnums to openapi
